\name{fac}
\alias{fac}
\docType{data}
\title{
factor "year"
}
\description{
factor "year" with seven levels, years 1974 to 1980
}
\usage{data(fac)}
\format{
  The format is:
 Factor with 7 levels "1974","1975",..,1980: 1 1 1 2 2 2 3 3 3 4 ...7
}
\source{
J.S. Pierre and C.A. Dedryver
}
\examples{
data(fac)
print(fac)
}
\keyword{datasets}
