\name{sit}
\Rdversion{1.1}
\alias{sit}
\docType{data}
\title{
Cereal aphids in Rennes
}
\description{
Maximum populations (log transformed) of the aphid \emph{sitobion avenae} in wheat fields from 1975 to 1981
}
\usage{data(sit)}
\format{
  The format is:
 num [1:7] 2.08e-01 2.12 3.36e+01 1.08 9.99e-16 ...
}
\source{
Pierre, J. S., Guillome, M. and Querrien, M. T. 1986. A Statistical and Graphic Method for Seeking in Which Periods of the Year Are 
the Animal Populations Peculiarly Sensitive to a Given Weather Component (Critical Periods of Time) - Application to the Case of Cereal Aphids. 
- Acta Oecologica-Oecologia Generalis 7: 365-380. (in french, english summary)
}
\examples{
data(sit)
## maybe str(sit) ; plot(sit) ...
}
\keyword{datasets}
