\name{time}
\Rdversion{1.1}
\alias{time}
\docType{data}
\title{
Fictitious daily temperatures 
}
\description{
Fictitious daily temperatures for 20 years, generated by \emph{rnorm}
}
\usage{data(time)}
\format{
  The format is:
 num [1:7300] 5.83 2.36 4.58 8.02 6.64 ...
}
\examples{
data(time)
## maybe str(time) ; plot(time) ...
}
\keyword{datasets}
