\name{time3}
\Rdversion{1.1}
\alias{time3}
\docType{data}
\title{
mean daily temperature data in Rennes (France)
}
\description{
mean daily temperature data in Rennes (France) from 1975 to 1981. february 29 of bissextile years discarded
}
\usage{data(time3)}
\format{
  The format is:
 num [1:2555] 1 2.3 1.7 6.1 1.9 5.2 8.9 10.5 8.4 6.3 ...
}
\source{
Pierre, J. S., Guillome, M. and Querrien, M. T. 1986. A Statistical and Graphic Method for Seeking in Which Periods of the Year Are 
the Animal Populations Peculiarly Sensitive to a Given Weather Component (Critical Periods of Time) - Application to the Case of Cereal Aphids. 
- Acta Oecologica-Oecologia Generalis 7: 365-380. (in french, english summary)
}
\examples{
data(time3)
## maybe str(time3) ; plot(time3) ...
}
\keyword{datasets}
