\name{wy}
\Rdversion{1.1}
\alias{wy}
\docType{data}
\title{
Series of fictitious observations
}
\description{
Series of fictitious observations linearly linked to a period of the series "time". Exact correlation.
}
\usage{data(wy)}
\format{
  The format is:
 num [1:20] 149 158 157 143 165 ...
}
\examples{
data(wy)
## maybe str(wy) ; plot(wy) ...
}
\keyword{datasets}
