\name{wy2}
\Rdversion{1.1}
\alias{wy2}
\docType{data}
\title{
Series of fictitious observations 
}
\description{
Series of fictitious observations linearly linked to a period of the series "time". incomplete correlation. Normal noise added
}
\usage{data(wy2)}
\format{
  The format is:
 num [1:20] 160 152 157 137 182 ...
}
\examples{
data(wy2)
## maybe str(wy2) ; plot(wy2) ...
}
\keyword{datasets}
