\name{fac}
\alias{fac}
\docType{data}
\title{
factor "year"
}
\description{
factor "year" with seven levels, years 1974 to 1980
}
\usage{data(fac)}
\format{
  The format is:
 Factor with 7 levels "1974","1975",..,1980: 1 1 1 2 2 2 3 3 3 4 ...7
}
\source{
Pierre, J. S., Guillome, M. and Querrien, M. T. 1986. A Statistical and Graphic Method for Seeking in Which Periods of the Year Are 
the Animal Populations Peculiarly Sensitive to a Given Weather Component (Critical Periods of Time) - Application to the Case of Cereal Aphids. 
- Acta Oecologica-Oecologia Generalis 7: 365-380. (in french, english summary)
}
\examples{
data(fac)
print(fac)
}
\keyword{datasets}
