\name{sit}
\Rdversion{1.1}
\alias{sit}
\docType{data}
\title{
Cereal aphids in Rennes
}
\description{
Maximum populations (log transformed) of the aphid \emph{sitobion avenae} in wheat fields from 1975 to 1981
}
\usage{data(sit)}
\format{
  The format is:
 num [1:7] 2.08e-01 2.12 3.36e+01 1.08 9.99e-16 ...
}
\source{
PIERRE J.S., GUILLOME M., QUERRIEN M.T., 1986. Une methode statistique et graphique de recherche des periodes de l'annee ou les populations animales sont particulierement sensibles a une composante donnee du climat (periodes critiques). Application au cas des pucerons des cereales. Acta Oecologica, Oecol. Gener., 7, 365-380.
}
\examples{
data(sit)
print(sit)
}
\keyword{datasets}
